images/api/requirements.txt
Flask==2.0.2
google-cloud-firestore==2.7.0
gunicorn==20.1.0
